+++ src/main/java/jm/security/example/service/RoleService.java
@@ -12,7 +12,7 @@ public interface RoleService {
 
     void update(Role role);
 
-    Role getById(int id);
+    Role getById(long id);
 
     Role getByName(String roleName);
