+++ src/main/java/jm/security/example/service/RoleServiceImpl.java
@@ -40,7 +40,7 @@ public class RoleServiceImpl implements RoleService{
 
     @Override
     @Transactional
-    public Role getById(int id) {
+    public Role getById(long id) {
         return roleDao.getById(id);
     }
